exported helper functions to seperate file
